+++ New_Extensibility_Model/Samples/VsProjectQueryAPISample/AddFile.cs
@@ -1,0 +1,112 @@
+﻿using System;
+using System.ComponentModel.Design;
+using System.Globalization;
+using System.Text;
+using System.Threading;
+using System.Threading.Tasks;
+using Microsoft.VisualStudio.ProjectSystem.Query;
+using System.Windows.Forms;
+using Microsoft.VisualStudio.Shell;
+using Microsoft.VisualStudio.Shell.Interop;
+using Task = System.Threading.Tasks.Task;
+
+namespace VsProjectQueryAPISample
+{
+	/// <summary>
+	/// Command handler
+	/// </summary>
+	internal sealed class AddFile
+	{
+		/// <summary>
+		/// Command ID.
+		/// </summary>
+		public const int CommandId = 0x0100;
+
+		/// <summary>
+		/// Command menu group (command set GUID).
+		/// </summary>
+		public static readonly Guid CommandSet = new Guid("0f208f00-10ab-4c3d-a89c-e8cbbbb9499f");
+
+		/// <summary>
+		/// VS Package that provides this command, not null.
+		/// </summary>
+		private readonly AsyncPackage package;
+
+		/// <summary>
+		/// Initializes a new instance of the <see cref="AddFile"/> class.
+		/// Adds our command handlers for menu (commands must exist in the command table file)
+		/// </summary>
+		/// <param name="package">Owner package, not null.</param>
+		/// <param name="commandService">Command service to add command to, not null.</param>
+		private AddFile(AsyncPackage package, OleMenuCommandService commandService)
+		{
+			this.package = package ?? throw new ArgumentNullException(nameof(package));
+			commandService = commandService ?? throw new ArgumentNullException(nameof(commandService));
+
+			var menuCommandID = new CommandID(CommandSet, CommandId);
+			var menuItem = new MenuCommand(this.Execute, menuCommandID);
+			commandService.AddCommand(menuItem);
+		}
+
+		/// <summary>
+		/// Gets the instance of the command.
+		/// </summary>
+		public static AddFile Instance
+		{
+			get;
+			private set;
+		}
+
+		/// <summary>
+		/// Gets the service provider from the owner package.
+		/// </summary>
+		private Microsoft.VisualStudio.Shell.IAsyncServiceProvider ServiceProvider
+		{
+			get
+			{
+				return this.package;
+			}
+		}
+
+		/// <summary>
+		/// Initializes the singleton instance of the command.
+		/// </summary>
+		/// <param name="package">Owner package, not null.</param>
+		public static async Task InitializeAsync(AsyncPackage package)
+		{
+			// Switch to the main thread - the call to AddCommand in AddFile's constructor requires
+			// the UI thread.
+			await ThreadHelper.JoinableTaskFactory.SwitchToMainThreadAsync(package.DisposalToken);
+
+			OleMenuCommandService commandService = await package.GetServiceAsync(typeof(IMenuCommandService)) as OleMenuCommandService;
+			Instance = new AddFile(package, commandService);
+		}
+
+		/// <summary>
+		/// This function is the callback used to execute the command when the menu item is clicked.
+		/// See the constructor to see how the menu item is associated with this function using
+		/// OleMenuCommandService service and MenuCommand class.
+		/// </summary>
+		/// <param name="sender">Event sender.</param>
+		/// <param name="e">Event args.</param>
+		private void Execute(object sender, EventArgs e)
+		{
+			ThreadHelper.ThrowIfNotOnUIThread();
+			StringBuilder message = new StringBuilder();
+			ThreadHelper.JoinableTaskFactory.Run(async () => {
+				var queryService = await this.package.GetServiceAsync<IProjectSystemQueryService, IProjectSystemQueryService>();
+				var querySpace = queryService.QueryableSpace;
+
+				var result = querySpace.Projects
+						 .Where(project => project.Name == "ConsoleApp1")
+						 .AsUpdatable()
+						 .CreateFile("CreatedFile.txt")
+				.ExecuteAsync();
+
+				message.Append("\n \n === New File Added to ConsoleApp1 === \n");
+
+				MessageBox.Show(message.ToString());
+			});
+		}
+	}
+}
